backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
google-generativeai
langchain
langchain-google-genai
